+++ src/main/java/com/example/service/InstructorAuthorizationService.java
@@ -3,7 +3,7 @@ package com.example.service;
 import com.example.model.Course;
 
 public interface InstructorAuthorizationService {
-    Course ensureAuthorizedCourseAccess(Integer instructorId, Integer courseId);
+    Course ensureAuthorizedCourseAccess(Long instructorId, Long courseId);
 }
